feat(member): /me 頁面整理 — 通知狀態 + 店名搬到頭像區
依實機 UI 回饋:
- 通知狀態 chip (🔔 已啟用通知) 移到頭像區跟「✓ 已綁定 LINE」並列
- 開啟通知按鈕 (PWA + 未訂閱時) 取代右上角「PWA 碼」位置
- 店名 (📍 信義店) 也加進頭像 chip 區,原 store info 區塊
  上面那個 standalone 店名標題改成「店家資訊」
- 拿掉 PWA 模式的「立即下單 / 逛商品 →」CTA 卡 (bar 已有「商品」tab)

底部 PushNotificationManager 卡簡化:
- 不再顯示「已啟用通知 / 尚未啟用」狀態列(已上移到頭像 chip)
- 留 重新連動 LINE + 推播說明 + debug 狀態

訂閱邏輯抽到 hook usePushNotification(jwt):
- /me 頁面層 lift state up,頭像 chip 跟底部 PushNotificationManager
  共用同一份 subscription state,避免 hook 跑兩次造成 double-subscribe
